+++ ar/edu/unlu/poo/estaciondeservicio/Empleado.java
@@ -10,7 +10,7 @@ public class Empleado {
     private String telefono;
     private ArrayList<Venta> ventasEmpleado = new ArrayList<>();
     //posiblemente eliminar esto
-    private Float totalVentasEmpleado = 0F;
+    private double totalVentasEmpleado = 0F;
 
     public void setNombre(String nombre) {
         this.nombre = nombre;
@@ -60,11 +60,11 @@ public class Empleado {
         return ventasEmpleado;
     }
 
-    public void setTotalVentasEmpleado(Float totalVentasEmpleado) {
+    public void setTotalVentasEmpleado(double totalVentasEmpleado) {
         this.totalVentasEmpleado = totalVentasEmpleado;
     }
 
-    public Float getTotalVentasEmpleado() {
+    public double getTotalVentasEmpleado() {
         return totalVentasEmpleado;
     }
 }
